refactor: Moves AttackResult to its own file
